% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensvm.grid.R
\name{gensvm.generate.cv.idx}
\alias{gensvm.generate.cv.idx}
\title{Generate a vector of cross-validation indices}
\usage{
gensvm.generate.cv.idx(n, folds)
}
\description{
This function generates a vector of length \code{n} with values from 0 to 
\code{folds-1} to mark train and test splits.
}

