% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensvm.refit.R
\name{gensvm.refit}
\alias{gensvm.refit}
\title{Train an already fitted model on new data}
\usage{
gensvm.refit(fit, X, y, p = NULL, lambda = NULL, kappa = NULL,
  epsilon = NULL, weights = NULL, kernel = NULL, gamma = NULL,
  coef = NULL, degree = NULL, kernel.eigen.cutoff = NULL,
  max.iter = NULL, verbose = NULL, random.seed = NULL)
}
\arguments{
\item{fit}{Fitted \code{gensvm} object}

\item{X}{Data matrix of the new data}

\item{y}{Label vector of the new data}

\item{verbose}{Turn on verbose output and fit progress. If NULL (the 
default) the value from the fitted model is chosen.}
}
\value{
a new fitted \code{gensvm} model
}
\examples{
x <- iris[, -5]
y <- iris[, 5]

# fit a standard model and refit with slightly different parameters
fit <- gensvm(x, y)
fit2 <- gensvm.refit(fit, x, y, epsilon=1e-8)

# refit a model returned by a grid search
grid <- gensvm.grid(x, y)
fit <- gensvm.refit(fit, x, y, epsilon=1e-8)

# refit on different data
idx <- runif(nrow(x)) > 0.5
x1 <- x[idx, ]
x2 <- x[!idx, ]
y1 <- y[idx]
y2 <- y[!idx]

fit1 <- gensvm(x1, y1)
fit2 <- gensvm.refit(fit1, x2, y2)

}
\author{
Gerrit J.J. van den Burg, Patrick J.F. Groenen \cr
Maintainer: Gerrit J.J. van den Burg <gertjanvandenburg@gmail.com>
}
\references{
Van den Burg, G.J.J. and Groenen, P.J.F. (2016). \emph{GenSVM: A Generalized 
Multiclass Support Vector Machine}, Journal of Machine Learning Research, 
17(225):1--42. URL \url{http://jmlr.org/papers/v17/14-526.html}.
}
\seealso{
\code{\link{gensvm}}, \code{\link{gensvm-package}}
}

