% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{gensvm.validate.param.grid}
\alias{gensvm.validate.param.grid}
\title{[internal] Validate parameter grid}
\usage{
gensvm.validate.param.grid(df)
}
\value{
TRUE if all values pass their respective conditions, FALSE 
otherwise.
}
\description{
Internal function to validate all parameters in a parameter 
grid.
}
\keyword{internal}
