% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{gensvm.validate.params}
\alias{gensvm.validate.params}
\title{[internal] Validate parameters}
\usage{
gensvm.validate.params(p = NULL, kappa = NULL, lambda = NULL,
  epsilon = NULL, gamma = NULL, weights = NULL, kernel = NULL, ...)
}
\value{
TRUE if all values pass their respective conditions, FALSE 
otherwise.
}
\description{
Internal function used to validate the parameters passed to the 
gensvm() function.
}
\keyword{internal}
