% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gensvm.R
\name{plot.gensvm}
\alias{plot.gensvm}
\title{Plot the simplex space of the fitted GenSVM model}
\usage{
\method{plot}{gensvm}(fit, y, x.test = NULL, with.margins = TRUE,
  with.shading = TRUE, with.legend = TRUE, center.plot = TRUE,
  xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{fit}{A fitted \code{gensvm} object}

\item{y}{the labels to color points with (if NULL the predicted labels are 
used)}

\item{with.margins}{plot the margins}

\item{with.shading}{show shaded areas for the class regions}

\item{with.legend}{show the legend for the class labels}

\item{center.plot}{ensure that the boundaries and margins are always visible 
in the plot}

\item{xlim}{allows the user to force certain plot limits. If set, these 
bounds will be used for the horizontal axis.}

\item{ylim}{allows the user to force certain plot limits. If set, these 
bounds will be used for the vertical axis and the value of center.plot will 
be ignored}

\item{...}{further arguments are passed to the builtin plot() function}

\item{x}{the dataset to plot (if NULL the training data is used)}
}
\value{
returns the object passed as input
}
\description{
This function creates a plot of the simplex space for a fitted 
GenSVM model and the given data set, as long as the dataset consists of only 
3 classes.  For more than 3 classes, the simplex space is too high 
dimensional to easily visualize.
}
\examples{
x <- iris[, -5]
y <- iris[, 5]

# train the model
fit <- gensvm(x, y)

# plot the simplex space
plot(fit)

# plot and use the true colors (easier to spot misclassified samples)
plot(fit, y)

# plot only misclassified samples
x.mis <- x[predict(fit) != y, ]
y.mis.true <- y[predict(fit) != y]
plot(fit, x.test=x.mis)
plot(fit, y.mis.true, x.test=x.mis)

}
\references{
Van den Burg, G.J.J. and Groenen, P.J.F. (2016). \emph{GenSVM: A Generalized 
Multiclass Support Vector Machine}, Journal of Machine Learning Research, 
17(225):1--42. URL \url{http://jmlr.org/papers/v17/14-526.html}.
}
\seealso{
\code{\link{plot.gensvm.grid}}, \code{\link{predict.gensvm}}, 
\code{\link{gensvm}}, \code{\link{gensvm-package}}
}
\author{
Gerrit J.J. van den Burg, Patrick J.F. Groenen \cr
Maintainer: Gerrit J.J. van den Burg <gertjanvandenburg@gmail.com>
}
